test(calendar-preview): group tests by subject, not by the pass that found them
23 files to 17. All 250 assertions unchanged; only their homes are.

`audit-fixed.test.tsx` collected findings under the number they were reported
as, and `regressions.test.tsx` grouped them by which audit pass found them.
Neither is a property of the component, so a CI failure named the incident
rather than the behaviour — `07: snapping never rolls into the next hour` told
you nothing about where to look. Both are distributed to the file that owns the
behaviour, and the numeric prefixes are gone.

  audit-fixed  -> input-nav 4, time-field 2, timezone 2, bounds-and-order 2,
                  keyboard 2, range-input 1, month-grid 1, calendar-preview 1
  regressions  -> calendar-preview 6, range-input 4, granularity 2,
                  timezone 1, input-nav 1, slots-documented 1

Four small files also folded into their subjects: `merge` into `input-nav`,
`month-sync` into `calendar-preview`, `data-slots` into `slots-documented`, and
`rerender` + `memo-stability` into `identity` — both of the latter ask what
survives a re-render, one about DOM nodes and one about a memo.

Verified by re-running all 14 mutation guarantees from their new homes, each
under the host timezone it needs: B1 5, B2 5, H12 1, B3 3, H9 1, B4 1, H3 4,
H2 1, H4 1, H5 1, H10 2, X2 1 failure respectively, plus H1 under
America/New_York and H11 under UTC. Those last two pass on any host whose zone
matches the display zone, which is the property they exist to expose.

Three things this move broke and the checks that did or did not catch them.
A `MONTH` collision between two merged files — caught by tsc. A stray copy of
`filter-chip.tsx` left in `components/calendar-preview/` by a careless restore
glob — caught by `exports.test.ts`, which flagged it as a type declared public
and absent from the barrel. And a `parseDate` test that a bad brace match
nested inside the `setTime` describe — caught by nothing: the count still read
250, no names duplicated, tsc was clean and the suite was green. Test count
parity is not evidence of a clean move.
