+++ packages/raystack/components/calendar-preview/__tests__/time-field.test.tsx
@@ -4,12 +4,14 @@ import { describe, expect, it, vi } from 'vitest';
 import { getSlot } from '~/test-utils/data-slots';
 import { CalendarPreview } from '../calendar-preview';
 import type { DateRangeValue } from '../calendar-preview-context';
-import { getHours, getMinutes } from '../date-adapter';
+import { getHours, getMinutes, isWithinTimeBounds } from '../date-adapter';
 
 const lastArg = (fn: { mock: { calls: unknown[][] } }) =>
   fn.mock.calls[fn.mock.calls.length - 1]?.[0];
 
 const hour = () => screen.getByLabelText('Hour');
+
+const MONTH = new Date(2024, 3, 1);
 const minute = () => screen.getByLabelText('Minute');
 
 const tree = (props: Record<string, unknown> = {}, fieldProps = {}) =>
@@ -145,5 +147,144 @@ describe('CalendarPreview.TimeField', () => {
     await user.keyboard('{Escape}');
     expect(hour()).toHaveValue('09');
     expect(onValueChange).not.toHaveBeenCalled();
+  });
+});
+
+/*
+ * Moved here from `audit-fixed.test.tsx`, which collected findings by the
+ * number they were reported under. Each assertion is unchanged; only its home
+ * is, so a failure lands beside the behaviour it describes.
+ */
+describe('regressions from the external audit', () => {
+  it('snapping never rolls into the next hour', async () => {
+    const user = userEvent.setup();
+    const onValueChange = vi.fn();
+    render(
+      <CalendarPreview
+        value={new Date(2024, 3, 17, 9, 0)}
+        onValueChange={onValueChange}
+      >
+        <CalendarPreview.TimeField step={15} />
+      </CalendarPreview>
+    );
+
+    const minute = screen.getByLabelText('Minute');
+    await user.clear(minute);
+    await user.type(minute, '59{Enter}');
+    const next = lastArg(onValueChange) as Date;
+    // `<= 59` was true of every Date ever constructed. The property is that the
+    // snap lands on the step grid without rolling the hour: 59 snaps to 45.
+    expect(next.getHours()).toBe(9);
+    expect(next.getMinutes()).toBe(45);
+    expect(next.getMinutes() % 15).toBe(0);
+  });
+
+  /*
+   * Asserting `isWithinBounds.length === 4` only checked the declared parameter
+   * count, and the case beside it was true in every zone — so making the
+   * function ignore `timeZone` entirely left all 33 tests here green, and all
+   * 358 across the repo. This asks the only question that separates the two:
+   * one instant that falls on different days depending on the zone it is read
+   * in, against a bound that sits between them.
+   */
+});
+
+describe('.TimeField honours the picker bounds', () => {
+  const setup = (props: Record<string, unknown>) => {
+    const onValueChange = vi.fn();
+    const onValidityChange = vi.fn();
+    render(
+      <CalendarPreview
+        defaultMonth={MONTH}
+        value={new Date(2024, 3, 17, 9, 0)}
+        onValueChange={onValueChange}
+        onValidityChange={onValidityChange}
+        {...props}
+      >
+        <CalendarPreview.TimeField />
+      </CalendarPreview>
+    );
+    return { onValueChange, onValidityChange };
+  };
+
+  it('refuses an hour past maxDate and reports why', async () => {
+    const user = userEvent.setup();
+    // Bounded at 10:00 *on the selected day*, so only a time comparison can
+    // catch this — `isWithinBounds` compares whole days and would pass it.
+    const { onValueChange, onValidityChange } = setup({
+      maxDate: new Date(2024, 3, 17, 10, 0)
+    });
+
+    await user.clear(screen.getByLabelText('Hour'));
+    await user.type(screen.getByLabelText('Hour'), '23{Enter}');
+
+    expect(onValueChange).not.toHaveBeenCalled();
+    expect(lastArg(onValidityChange)).toEqual({
+      valid: false,
+      reason: 'out-of-bounds'
+    });
+  });
+
+  it('refuses an hour before minDate', async () => {
+    const user = userEvent.setup();
+    const { onValueChange, onValidityChange } = setup({
+      minDate: new Date(2024, 3, 17, 8, 0)
+    });
+
+    await user.clear(screen.getByLabelText('Hour'));
+    await user.type(screen.getByLabelText('Hour'), '07{Enter}');
+
+    expect(onValueChange).not.toHaveBeenCalled();
+    expect(lastArg(onValidityChange)).toEqual({
+      valid: false,
+      reason: 'out-of-bounds'
+    });
+  });
+
+  it('leaves the whole last day usable under a day-level maxDate', async () => {
+    const user = userEvent.setup();
+    // The ordinary way a picker is bounded: a plain day, at midnight. Read
+    // literally as an instant it would forbid every time on the 17th, which
+    // is not what it means anywhere else in the component.
+    const { onValueChange, onValidityChange } = setup({
+      maxDate: new Date(2024, 3, 17)
+    });
+
+    await user.clear(screen.getByLabelText('Hour'));
+    await user.type(screen.getByLabelText('Hour'), '10{Enter}');
+
+    expect(getHours(lastArg(onValueChange) as Date)).toBe(10);
+    expect(lastArg(onValidityChange)).toEqual({ valid: true });
+  });
+
+  it('still rejects the day after a day-level maxDate', () => {
+    // The day bound has not gone soft — it is applied first, inclusive.
+    expect(
+      isWithinTimeBounds(
+        new Date(2024, 3, 18, 9, 0),
+        undefined,
+        new Date(2024, 3, 17)
+      )
+    ).toBe(false);
+    expect(
+      isWithinTimeBounds(
+        new Date(2024, 3, 17, 23, 59),
+        undefined,
+        new Date(2024, 3, 17)
+      )
+    ).toBe(true);
+  });
+
+  it('commits an in-bounds hour and reports valid', async () => {
+    const user = userEvent.setup();
+    const { onValueChange, onValidityChange } = setup({
+      maxDate: new Date(2024, 3, 17, 10, 0)
+    });
+
+    await user.clear(screen.getByLabelText('Hour'));
+    await user.type(screen.getByLabelText('Hour'), '10{Enter}');
+
+    expect(getHours(lastArg(onValueChange) as Date)).toBe(10);
+    expect(lastArg(onValidityChange)).toEqual({ valid: true });
   });
 });
